pyproject.toml
[tool.poetry]
name = "does"
version = "0.1.0"
description = ""
authors = ["André Carvalho <alac1984@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.1"
beautifulsoup4 = "^4.13.3"
lxml = "^5.3.1"
pypdf = "^5.3.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
sqlmodel = "^0.0.23"
alembic = "^1.14.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"


[tool.poetry.group.dev.dependencies]
pynvim = "^0.5.2"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
mypy = "^1.15.0"
flake8 = "^7.1.2"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
